+++ services/importer/spec/acceptance/geojson_spec.rb
@@ -1,5 +1,6 @@
 # encoding: utf-8
 require_relative '../../../../spec/rspec_configuration'
+require_relative '../../../../spec/spec_helper'
 require 'rspec/core'
 require 'rspec/expectations'
 require 'rspec/mocks'
@@ -15,20 +16,19 @@ require_relative 'acceptance_helpers'
 require_relative 'no_stats_context'
 
 
-include CartoDB::Importer2
 
 describe 'geojson regression tests' do
   include AcceptanceHelpers
   include_context "no stats"
 
   before do
-    @pg_options  = Factories::PGConnection.new.pg_options
+    @pg_options  = ::CartoDB::Importer2::Factories::PGConnection.new.pg_options
   end
 
   it 'imports a file exported from CartoDB' do
     filepath    = path_to('tm_world_borders_simpl_0_8.geojson')
-    downloader  = Downloader.new(filepath)
-    runner      = Runner.new({
+    downloader  = ::CartoDB::Importer2::Downloader.new(filepath)
+    runner      = ::CartoDB::Importer2::Runner.new({
                                pg: @pg_options,
                                downloader: downloader,
                                log: CartoDB::Importer2::Doubles::Log.new,
@@ -40,8 +40,8 @@ describe 'geojson regression tests' do
   it 'imports a file from a url with params' do
     filepath    = 'https://raw.github.com/benbalter/dc-wifi-social/master' +
                   '/bars.geojson?foo=bar'
-    downloader  = Downloader.new(filepath)
-    runner      = Runner.new({
+    downloader  = ::CartoDB::Importer2::Downloader.new(filepath)
+    runner      = ::CartoDB::Importer2::Runner.new({
                                pg: @pg_options,
                                downloader: downloader,
                                log: CartoDB::Importer2::Doubles::Log.new,
@@ -52,8 +52,8 @@ describe 'geojson regression tests' do
 
   it "raises if GeoJSON isn't valid" do
     filepath    = path_to('invalid.geojson')
-    downloader  = Downloader.new(filepath)
-    runner      = Runner.new({
+    downloader  = ::CartoDB::Importer2::Downloader.new(filepath)
+    runner      = ::CartoDB::Importer2::Runner.new({
                                pg: @pg_options,
                                downloader: downloader,
                                log: CartoDB::Importer2::Doubles::Log.new,