+++ services/importer/spec/acceptance/zip_spec.rb
@@ -1,5 +1,7 @@
 # encoding: utf-8
 require_relative '../../../../spec/rspec_configuration'
+require_relative '../../../../spec/spec_helper'
+require_relative '../../../../spec/spec_helper'
 require_relative '../../lib/importer/runner'
 require_relative '../../lib/importer/job'
 require_relative '../../lib/importer/downloader'
@@ -10,21 +12,19 @@ require_relative 'acceptance_helpers'
 require_relative 'cdb_importer_context'
 require_relative 'no_stats_context'
 
-include CartoDB::Importer2
-
 describe 'zip regression tests' do
   include AcceptanceHelpers
   include_context "cdb_importer schema"
   include_context "no stats"
 
   before do
-    @pg_options  = Factories::PGConnection.new.pg_options
+    @pg_options  = ::CartoDB::Importer2::Factories::PGConnection.new.pg_options
   end
 
   it 'returns empty results if no supported files in the bundle' do
     filepath    = path_to('one_unsupported.zip')
-    downloader  = Downloader.new(filepath)
-    runner      = Runner.new({
+    downloader  = ::CartoDB::Importer2::Downloader.new(filepath)
+    runner      = ::CartoDB::Importer2::Runner.new({
                                pg: @pg_options,
                                downloader: downloader,
                                log: CartoDB::Importer2::Doubles::Log.new,
@@ -37,8 +37,8 @@ describe 'zip regression tests' do
 
   it 'ignores unsupported files in the bundle' do
     filepath    = path_to('one_unsupported_one_valid.zip')
-    downloader  = Downloader.new(filepath)
-    runner      = Runner.new({
+    downloader  = ::CartoDB::Importer2::Downloader.new(filepath)
+    runner      = ::CartoDB::Importer2::Runner.new({
                                pg: @pg_options,
                                downloader: downloader,
                                log: CartoDB::Importer2::Doubles::Log.new,
@@ -51,8 +51,8 @@ describe 'zip regression tests' do
 
   it 'imports a zip with >1 file successfully' do
     filepath    = path_to('multiple_csvs.zip')
-    downloader  = Downloader.new(filepath)
-    runner      = Runner.new({
+    downloader  = ::CartoDB::Importer2::Downloader.new(filepath)
+    runner      = ::CartoDB::Importer2::Runner.new({
                                pg: @pg_options,
                                downloader: downloader,
                                log: CartoDB::Importer2::Doubles::Log.new,
@@ -70,8 +70,8 @@ describe 'zip regression tests' do
 
   it 'imports a maximum of Runner::MAX_TABLES_PER_IMPORT files from a zip, but doesnt errors' do
     filepath    = path_to('more_than_10_files.zip')
-    downloader  = Downloader.new(filepath)
-    runner      = Runner.new({
+    downloader  = ::CartoDB::Importer2::Downloader.new(filepath)
+    runner      = ::CartoDB::Importer2::Runner.new({
                                pg: @pg_options,
                                downloader: downloader,
                                log: CartoDB::Importer2::Doubles::Log.new,
@@ -90,8 +90,8 @@ describe 'zip regression tests' do
   it 'imports a shapefile that includes a xxx.VERSION.txt file skipping it' do
     # http://www.naturalearthdata.com/downloads/
     filepath    = path_to('shapefile_with_version_txt.zip')
-    downloader  = Downloader.new(filepath)
-    runner      = Runner.new({
+    downloader  = ::CartoDB::Importer2::Downloader.new(filepath)
+    runner      = ::CartoDB::Importer2::Runner.new({
                                pg: @pg_options,
                                downloader: downloader,
                                log: CartoDB::Importer2::Doubles::Log.new,
@@ -109,8 +109,8 @@ describe 'zip regression tests' do
 
     it 'imports all non-failing items from a zip without failing the whole import' do
     filepath    = path_to('file_ok_and_file_ko.zip')
-    downloader  = Downloader.new(filepath)
-    runner      = Runner.new({
+    downloader  = ::CartoDB::Importer2::Downloader.new(filepath)
+    runner      = ::CartoDB::Importer2::Runner.new({
                                pg: @pg_options,
                                downloader: downloader,
                                log: CartoDB::Importer2::Doubles::Log.new,