+++ spec/factories/visualizations.rb
@@ -1,3 +1,5 @@
+require 'uuidtools'
+
 FactoryGirl.define do
 
   random = UUIDTools::UUID.timestamp_create.to_s
@@ -14,4 +16,10 @@ FactoryGirl.define do
     privacy 'public'
   end
 
+  factory :carto_visualization, class: Carto::Visualization do
+    id { UUIDTools::UUID.random_create.to_s }
+    type 'derived'
+    name 'factory visualization'
+    privacy 'public'
+  end
 end