+++ NEWS.md
@@ -1,7 +1,10 @@
 3.12.2 (2016-05-15)
 -------------------
+* Support for HTTP Header authentication.
 * Update navigation in public pages.
 * New feature in organizations management: user quota slider.
+* Added `data_library` to the config. Usage instructions:
+  1. Configure data-library path. You can copy `data_library` config from `config/app_config.yml.sample` into your `config/app_config.yml`.
 
 3.12.0 (2016-01-05)
 -------------------